% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{blockLengths}
\alias{blockLengths}
\title{Find the number of possible permutations for each block.}
\usage{
blockLengths(permBlocks)
}
\arguments{
\item{permBlocks}{A nested list containing the possible permutation
blocks generated by \code{\link{permutationBlocksByCol}}.}
}
\description{
This uses the output of \code{\link{permutationBlocksByCol}} and finds the
number of permuatations in each permutation block.
}

