% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{checkPermutedDesign}
\alias{checkPermutedDesign}
\title{Identify which main effect a permutation is associated with}
\usage{
checkPermutedDesign(X, perm)
}
\arguments{
\item{X}{A matrix.}

\item{perm}{A vector of integers denoting a permutation of design points.}
}
\value{
A vector of integers
}
\description{
This function looks at a matrix and determines which columns are not changed
by a permutation. TRUE indicates that the column values are unchanged.
}

