% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{columnLevels}
\alias{columnLevels}
\title{Return the levels in a column of a matrix}
\usage{
columnLevels(X, col)
}
\arguments{
\item{X}{A matrix.}

\item{col}{A number indicating the column.}
}
\description{
This returns the unique values in a specified column of a matrix.
}
\details{
This is used by \code{\link{permutableIndicesByCol}}.
}

