% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_SSM.R
\name{compute.covariance.from.distance}
\alias{compute.covariance.from.distance}
\title{Compute unscaled covariance matrix from a supplied distance matrix and length
parameter.}
\usage{
compute.covariance.from.distance(distance, r, type = "exp", ssm)
}
\arguments{
\item{distance}{A matrix. A symmetric matrix of distances between design
points.}

\item{r}{A number. The length parameter for the correlation function}

\item{type}{(optional) Character. Specifies the correlation function. Default
behaviour is to use \code{"exp"} for a squared exponential correlation.
\code{"Matern32"} uses a Matern 3/2 correlation.}

\item{ssm}{(obsolete) An SSM object. Used for the removed
\code{"ssm"} correlation function which tried to use the local smoothness
of the SSM to measure distances.}
}
\description{
This is a legacy function and is not used.
}

