% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_SSM.R
\name{compute.covariance}
\alias{compute.covariance}
\title{Compute the unscaled covariance matrix.}
\usage{
compute.covariance(ssm, r)
}
\arguments{
\item{ssm}{An SSM object.}

\item{r}{A number. The length parameter used by the correlation function.}
}
\value{
The unscaled covariance matrix.
}
\description{
Computes the unscaled covariance matrix for an SSM object for a given
length parameter.
}
\details{
The covariance matrix uses the correlation function specified by the SSM slot
\code{type}.  \code{"matern32"} uses a Matern 3/2 correlation while
\code{"exp"} uses a squared exponential.
}

