% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_SSM.R
\name{compute.interactions}
\alias{compute.interactions}
\title{Compute Total interaction indices and Sobol indices for higher order
interactions.}
\usage{
compute.interactions(ssm)
}
\arguments{
\item{ssm}{An SSM object.}
}
\value{
An SSM object.  This is the \code{ssm} input with the
relevant slots updated.
}
\description{
This function computes the Total interaction indices of all second order
interactions present in the model.  If \eqn{d < 11} then it also computes the
Sobol indices of all interactions (second order and higher) in the model. It
identifies the relevant rows of the SSM \code{basis} slot for each
interaction and sums the term variances. Called by
\code{\link{update_sensitivity}}.
}

