% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SSM.R
\name{compute.residuals}
\alias{compute.residuals}
\title{Compute the Leave-One-Out error at all design points.}
\usage{
compute.residuals(ssm)
}
\arguments{
\item{ssm}{An SSM object that has valid entries for the slots
\code{response}, \code{K}, and \code{design_model_matrix}.}
}
\value{
A vector of Leave-One_Out errors.
}
\description{
This function repeatedly fits the SSM to all but one held out point in turn
and computes the prediction error at the held out point.  These errors are
collated into a vector indicating the Leave-One-Out error at each design
point.
}
\keyword{internal}

