% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_SSM.R
\name{compute.total.effects}
\alias{compute.total.effects}
\title{Compute Total effects}
\usage{
compute.total.effects(ssm)
}
\arguments{
\item{ssm}{An SSM object.}
}
\value{
An SSM object.  This is the \code{ssm} input with the
relevant slots updated.
}
\description{
This function computes the Total indices of the model variables.  It
identifies the relevant rows of the SSM \code{basis} slot and sums the term
variances. Called by \code{\link{update.sensitivity}}.
}

