% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SSM.R
\name{construct.P}
\alias{construct.P}
\title{Construct the change of basis matrix from multivariate monomials to Legendre
polynomials.}
\usage{
construct.P(basis)
}
\arguments{
\item{basis}{A matrix. Rows are taken as the degree of the associated
monomial.}
}
\value{
A matrix which functions as a change of basis from monomials to
  Legendre.
}
\description{
Let \eqn{f(x)} be the vector of monomials implied by the matrix \code{basis}.
Let \eqn{P} be the matrix generated by \code{Legendre_P}.  Then
\eqn{g(x)=P^Tf(x)} is the vector of Legendre polynomials with leading terms
corresponding to \code{basis}.
}

