% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{constructDelta}
\alias{constructDelta}
\title{Construct the full \eqn{\Delta} matrix.}
\usage{
constructDelta(ssm)
}
\arguments{
\item{ssm}{An SSM class object.}
}
\value{
A vector containing the diagonal entries of the \eqn{\Delta} matrix.
}
\description{
This constructs the diagonal \eqn{\Delta} matrix that contains the
normalisation constants for the basis polynomials.
}
\details{
This assumes that the basis is formed from Legendre polynomials and is not
appropriate if the basis is user-defined.
}

