% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{constructH}
\alias{constructH}
\title{Construct the H matrix defining the indicator polynomials used in the
construction of an SSM.}
\usage{
constructH(ssm, type = 1)
}
\arguments{
\item{ssm}{An SSM class object.}

\item{type}{(optional) Determines the way \eqn{H} is computed. The default
\code{type = 1} will construct and invert C, taking the appropriate
sub-matrix. However, \code{type = 2} will use submatrices of the design
model matrix and the K matrix which may or may not be quicker (to be
determined through testing).}
}
\value{
The H matrix.
}
\description{
Given an SSM this function computes the H matrix that defines a size
\eqn{n} basis of the model such that \eqn{Hy=\theta}.
}
\details{
The method of computation when the \code{type} argument is set to 2 assumes
that the first \eqn{d+1} columns of the design model matrix are associated
with the affine polynomials in the model basis.  Therefore this method may
not be appropriate if a user-defined basis has been used and \code{type}
should be set to 1.
}

