% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{constructQ}
\alias{constructQ}
\title{Construct the Q matrix for an effect or interaction in an SSM.}
\usage{
constructQ(ssm, indices, type = 1)
}
\arguments{
\item{ssm}{An SSM class object.}

\item{indices}{A vector of integers identifying the effect or interaction of
interest.  For example, \code{indices = c(1, 2)} indicates the interaction
between \eqn{x_1} and \eqn{x_2} while \code{indices = 1} indicates the main
effect of \eqn{x_1}. If set to zero, the output matrix will define the
total variance.}

\item{type}{(optional) This is passed to \code{\link{constructH}} and
determines the method of computing the \eqn{H} matrix.  Must be 1 or 2.
used to construct \eqn{Q}. See \code{\link{constructH}} for details.}
}
\value{
The Q matrix used in the variance formulation \eqn{y^TQy}.
}
\description{
This function takes an ssm as an input along with a vector of integers that
denotes the effect or interaction of interest.  It then computes the relevant
Q matrix that is used to compute the FANOVA variance of that
effect/interaction using the formulation \eqn{y^TQy}.
}

