% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{createPermFromBlocks}
\alias{createPermFromBlocks}
\title{Given an integer, select the appropriate permutation blocks}
\usage{
createPermFromBlocks(ident, permNumbers, vec = NULL)
}
\arguments{
\item{ident}{An integer.}

\item{permNumbers}{A vector of integers indicating the number of
permutations in each block.}

\item{vec}{This argument holds the vector containing the selected
permutations in each block, hence it being set to NULL at first.}
}
\description{
This is a recursive function used to select the permutation blocks uniquely
identified by an integer between one and the total number of allowable
permutations.
}

