% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{createPermFromNumber}
\alias{createPermFromNumber}
\title{Creates the permutation of vector uniquely defined by a number}
\usage{
createPermFromNumber(number, inp, out = NULL)
}
\arguments{
\item{number}{An integer. The cycle of permutations repeats for integers
higher than the total number of possible permutations.}

\item{inp}{A vector}

\item{out}{This argument holds the permutation as it is constructed so it
needs to be set to NULL for the initial call.}
}
\description{
Every permutation of a vector can be assigned a unique number between one and
the total number of possible permutations.  This function supplies the
permutation associated with the given number. It does this recursively.
}

