% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{createPermFromNumbersAndBlocks}
\alias{createPermFromNumbersAndBlocks}
\title{Creates the permutation corresponding to a set of sub-permutations}
\usage{
createPermFromNumbersAndBlocks(perm, block)
}
\arguments{
\item{perm}{A vector of integers}

\item{block}{A list of integer vectors specifying the numbers to permute in
each subset.}
}
\description{
Creates the permutation corresponding to a set of sub-permutations
}

