% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{createPermutation}
\alias{createPermutation}
\title{Create a permutation that only permutes inside the allowed blocks}
\usage{
createPermutation(permblocks, bl)
}
\arguments{
\item{bl}{A vector usually generated by \code{\link{blockLengths}} that
gives the number of permutations in each block.}

\item{permBlocks}{A nested list containing the possible permutation
blocks generated by \code{\link{permutationBlocksByCol}}.}
}
\value{
A vector indicating a row permutation.
}
\description{
This function operates on the permutation blocks list created by
\code{\link{permutationBlocksByCol}} and the vector of block lengths created
by \code{\link{blockLengths}}.  It is used by
\code{\link{createPermutationByCol}}.
}

