% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{createPermutationByCol}
\alias{createPermutationByCol}
\title{Create n permutations that do not change the i-th column of a matrix}
\usage{
createPermutationByCol(X, i, n, exit = 10 * n)
}
\arguments{
\item{X}{A matrix}

\item{i}{The column to remain unchanged}

\item{n}{The number of permutations desired.}

\item{exit}{(optional) If this many permutations are generated in a row
without finding a new one, the function will exit and return the
permutations already generated.}
}
\value{
A list of permutations
}
\description{
WARNING - this uses a random sampling to generate permutations so it can be
inefficient. Once many permutations have been generated, the chance of
a sample being a new permutation decreases.  As a check, the generation will
end if \code{exit} permutations are generated in a row without finding a new
one.
}

