% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{createPermutationFromP}
\alias{createPermutationFromP}
\title{A variation of createPermutation that uses a generated vector}
\usage{
createPermutationFromP(p, permblocks)
}
\arguments{
\item{p}{A vector of integers indicating the block permutations to use.}

\item{permblocks}{A nested list containing the possible permutation
blocks generated by \code{\link{permutationBlocksByCol}}.}
}
\description{
This is used by \code{\link{generateAllPerms}} to create all permutations.
}

