% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{generateAllPerms}
\alias{generateAllPerms}
\title{Generate all possible permutations of a design}
\usage{
generateAllPerms(permBlocks, bl)
}
\arguments{
\item{permBlocks}{A nested list containing the possible permutation
blocks generated by \code{\link{permutationBlocksByCol}}.}

\item{bl}{A vector of block lengths}
}
\description{
This generates all possible permutations of each block, then combines them
all to construct all possible permutations.
}

