% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{generatePermutations}
\alias{generatePermutations}
\title{Generate n permutations that do not change the given column of a matrix}
\usage{
generatePermutations(n, X, col)
}
\arguments{
\item{n}{The number of desired permutations. If this is greater than the
total number of allowable permutations, only the allowable permutations
will be output.}

\item{X}{A matrix}

\item{col}{The index of the column that must remain unchanged.}
}
\description{
Generate n permutations that do not change the given column of a matrix
}

