% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SSM.R
\name{get.K.element}
\alias{get.K.element}
\title{Compute entry of K matrix.}
\usage{
get.K.element(n, m, basis)
}
\arguments{
\item{n}{A number. Identifies the first basis element in the inner product.}

\item{m}{A number. Identifies the second basis element in the inner product.}

\item{basis}{A matrix. Rows are taken as exponent vectors.}
}
\value{
The inner product of the \eqn{n, m}-th basis elements as defined for
  \code{\link{construct.K}}.
}
\description{
Called by \code{\link{construct.K}} to compute entries of the K matrix.
}

