% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{getPermutationFromNumber}
\alias{getPermutationFromNumber}
\title{Get nth permutation from a permutation block.}
\usage{
getPermutationFromNumber(permBlock, n)
}
\arguments{
\item{permBlock}{A single permutation block.}

\item{n}{An integer.}
}
\description{
Get nth permutation from a permutation block.
}

