% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_SSM.R
\name{identify.main.effect.terms}
\alias{identify.main.effect.terms}
\title{Identify main effect terms}
\usage{
\method{identify}{main.effect.terms}(i, basis)
}
\arguments{
\item{i}{A number indicating the main effect of interest.}

\item{basis}{A matrix.  A matrix where each row is an exponent vector of a
monomial basis.}
}
\value{
A logical vector.  Each term corresponds with the associated row of
\code{basis}.
}
\description{
This is used by the sensitivity computation functions to identify which basis
terms are associated with each main effect.
}

