% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_table.r
\docType{data}
\name{input_table}
\alias{input_table}
\title{Example input table containing sewage treatment plants and lakes}
\format{
An object of class data.frame with 683 rows and 29 columns, of which the following columns are essential to be used with function \code{\link[=wrap_table]{wrap_table()}}:
\itemize{
\item ID
\item inhabitants
\item ID_next
\item Q347_L_s_min
\item Q347_L_s_max
}

The follwing columns are required if \verb{in [wrap_table()] is set to}:
\itemize{
\item nitrification
\item denitrification
\item P_elimination
\item type_advanced_treatment
\item starting_year_advanced_treatment
\item redirecting_STP_target_STP_ID
}

The follwing columns are required if \verb{in [wrap_table()] is set to}:
\itemize{
\item lake_elimination_min
\item lake_elimination_max
}

The follwing columns are required if \verb{in [wrap_table()] is set to}:
\itemize{
\item STP_elimination_min
\item STP_elimination_max
\item 
}

The follwing column is required if \verb{in [wrap_table()] is set to}:
\itemize{
\item additional_absolut_load
}
}
\usage{
See function [wrap_table()].
}
\description{
This table is mainly intended as an example \code{input_table} to be used with function \code{\link[=wrap_table]{wrap_table()}}.
It contains information on sewage treatment plants (STPs) from Switzerland and neighbouring countries, as well as swiss lakes.
}
\author{
Rebekka Gulde \email{rebekka.gulde@vsa.ch}
}
\keyword{datasets}
