% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{levelIndices}
\alias{levelIndices}
\title{Identify the positions of a given value in a vector}
\usage{
levelIndices(i, vec)
}
\arguments{
\item{i}{A value to search for.}

\item{vec}{The vector to search.}
}
\description{
This is used by \code{\link{permutableIndicesByCol}}.
}

