% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_SSM.R
\name{lineij}
\alias{lineij}
\title{Average the values in a vector between two cutoff points specified by a
separate vector.}
\usage{
lineij(x, smoothness, smoothnessdesign)
}
\arguments{
\item{x}{A vector of length two. Specifies the endpoints of the desired
interval of integration.}

\item{smoothness}{A numeric vector.}

\item{smoothnessdesign}{A numeric vector of the same length as
\code{smoothness}.}
}
\value{
A number.
}
\description{
The values of \code{smoothnessdesign} that are in the interval specifed by
\code{x} are identified and the associated values in \code{smoothness} are
averaged. This is used in the computation of line integrals of smoothness
used by a particular distance measure when computing metamodel error
estimates.
}
\keyword{internal}

