% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_topology.r
\name{make_topology}
\alias{make_topology}
\title{Calculate relational topology matrix for STP and lake nodes in a river network.}
\usage{
make_topology(
  ID = FALSE,
  ID_next,
  NA_next_ignore = FALSE,
  insert_id_in_topo_matrix = FALSE,
  only_direct_upstream = FALSE
)
}
\arguments{
\item{ID}{Character or integer vector containing the IDs of river network nodes such as sewage treatment plants (STPs), lakes or similar.}

\item{ID_next}{Character or integer vector of length equal to ID, containing the ID of the next node directly downstream of each node (if any, else set entry to NA).}

\item{NA_next_ignore}{Logical. Set entries in vector ID_next which are not found in vector ID to NA? Else, throw an error (the default, FALSE).}

\item{insert_id_in_topo_matrix}{Logical, default FALSE. Insert IDs into returned matrix.}

\item{only_direct_upstream}{Logical, default FALSE. Return only STPs/lakes directly upstream?}
}
\value{
A matrix with dimensions and column/row names equal to argument ID. Columns contain 1
(or IDs for insert_id_in_topo_matrix) to denote upstream nodes of each STP/lake node, and 0 otherwise.
}
\description{
Calculates a topology matrix to mark all sewage treatment plants (STPs) or lakes (or similar network nodes) upstream
of an individual STP/lake in a river network. Using only_direct_upstream, these can either be all nodes upstream, or only those in direct
upstream adjacency. Mainly to be used in function calc_load().
}
\examples{
ID <- c(1, 2, 3, 4, 5)
ID_next <- c(4, 3, 4, 5, NA) # node with ID = 5 has no downstream node (hence ID_next = NA), but all nodes discharge to it (hence all its column set to 1):
make_topology(ID, ID_next)
}
\seealso{
{\code{\link[=calc_load]{calc_load()}}}
}
