% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_SSM.R
\name{optimize.by.interval.maximum}
\alias{optimize.by.interval.maximum}
\title{Optimize concentrated likelihood.}
\usage{
optimize.by.interval.maximum(int, ...)
}
\arguments{
\item{int}{The maximum endpoint of the interval of optimization.}
}
\description{
Wrapper for optimizing \code{concentrated.likelihood}. Used by
\code{\link{estimate.GP}}.
}

