% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{origPerm}
\alias{origPerm}
\title{Get the identity permutation from the i-th block}
\usage{
origPerm(i, permBlocks)
}
\arguments{
\item{i}{A number.}

\item{permBlocks}{A nested list containing the possible permutation
blocks generated by \code{\link{permutationBlocksByCol}}.}
}
\description{
This is used by \code{\link{createPermutationByCol}} to put the generated
permutation in the correct order.
}

