% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_SSM.R
\name{partial.deriv.ssm}
\alias{partial.deriv.ssm}
\title{Compute second partial derivative of a smooth supersaturated model at all
design points.}
\usage{
partial.deriv.ssm(indices, ssm)
}
\arguments{
\item{indices}{A vector of integers specifying the two variables by which we
take the partial derivatives with respect to.}

\item{ssm}{An SSM object.}
}
\value{
A single column vector containing the second partial derivatives with
respect to the requested variables, evaluated at all design points.
}
\description{
Computes a second partial derivative of a smooth supersaturated model at a
design point. Used in the computation of distance measures based on local
smoothness.
}

