% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{permutableIndicesByCol}
\alias{permutableIndicesByCol}
\title{Identify permutable matrix rows by column.}
\usage{
permutableIndicesByCol(X, col)
}
\arguments{
\item{X}{A matrix.}

\item{col}{A number indicating the column that should remain unchanged.}
}
\description{
This returns a list where each list entry is a set of matrix rows that can
be permuted without changing the column denoted by \code{col}.
}

