% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{permutationBlocksByCol}
\alias{permutationBlocksByCol}
\title{Generate possible permutation blocks for a matrix column}
\usage{
permutationBlocksByCol(X, col)
}
\arguments{
\item{X}{A matrix.}

\item{col}{A number indicating the column that is to be unchanged by
permutation.}
}
\description{
This generates a nested list containing all possible permutation blocks that
leave a specified column unchanged.
}

