% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{permute}
\alias{permute}
\title{permute a vector}
\usage{
permute(perm, vec)
}
\arguments{
\item{perm}{A permutation held as a vector of integers.}

\item{vec}{A vector (must be the same length as \code{perm}).}
}
\description{
permute a vector
}

