% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{randomOneToN}
\alias{randomOneToN}
\title{Generate single random integer between 1 and N}
\usage{
randomOneToN(N)
}
\arguments{
\item{N}{An integer.}
}
\description{
Generate single random integer between 1 and N
}

