% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSM_method.R
\docType{methods}
\name{show,SSM-method}
\alias{show,SSM-method}
\title{Summarise SSM class object}
\usage{
\S4method{show}{SSM}(object)
}
\arguments{
\item{SSM}{An SSM object}
}
\description{
Printing an SSM will summarise the parameters d, N, and n.  If no model has
been fit then this will be noted, otherwise The smoothness of the SSM will be
shown.  If sensitivity analysis has been performed, the Sobol indices and
Total indices for main effects are displayed and if cross-validation has been
performed the Standardised LOO RMSE is also shown.
}

