% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_SSM.R
\name{smoothness.over.design}
\alias{smoothness.over.design}
\title{Compute the smoothness of an SSM at all design points.}
\usage{
smoothness.over.design(ssm)
}
\arguments{
\item{ssm}{An SSM object.}
}
\value{
A vector of integers containing the smoothness at each design point.
}
\description{
This function evaluates the Frobenius norm of the Hessian matrix at all
design points. Used in the compuation of distance measures based on local
smoothness.
}

