% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_patch_baseline}
\alias{ssm_get_patch_baseline}
\title{Retrieves information about a patch baseline}
\usage{
ssm_get_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to retrieve.}
}
\description{
Retrieves information about a patch baseline.
}
\section{Request syntax}{
\preformatted{ssm$get_patch_baseline(
  BaselineId = "string"
)
}
}

\keyword{internal}
