\name{SSM-package}
\alias{SSM-package}
\alias{SSM}
\docType{package}
\title{
\packageTitle{SSM}
}
\description{
\packageDescription{SSM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SSM}
\packageIndices{SSM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SSM}

Maintainer: \packageMaintainer{SSM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
