% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_table.r
\name{wrap_table}
\alias{wrap_table}
\title{Wrapper to calculate loads from pre-defined table input}
\usage{
wrap_table(
  model_input_table = NULL,
  STP_scenario_year = as.numeric(strsplit(as.character(Sys.Date()), "-")[[1]][1]),
  STP_reroute = TRUE,
  STP_filter_steps = TRUE,
  STP_discharge_per_capita = 400,
  compound_name = "not_specified",
  scenario_name = compound_name,
  compound_load_gramm_per_capita_and_day,
  compound_elimination_STP = NULL,
  compound_elimination_method = "micropollutants",
  with_lake_elimination = FALSE,
  add_absolute_load = FALSE,
  use_columns_local_discharge = c("Q347_L_s_min", "Q347_L_s_max"),
  use_STP_elimination_rate = c("STP_elimination_min", "STP_elimination_max"),
  add_columns_from_model_input_table = c("ID_next", "X_position", "Y_position"),
  path_out = FALSE,
  overwrite = TRUE,
  write_csv = TRUE,
  use_sep_csv = " "
)
}
\value{
An integer vector the same length as \code{x}.
\code{NA} strings have \code{NA} length.
}
\description{
asd
}
\examples{
Bla.

}
\seealso{
\code{\link[=calc_load]{calc_load()}}
}
