% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_permutations.R
\name{yQy}
\alias{yQy}
\title{Compute \eqn{y'Qiy \ y'Qy} for a given permutation}
\usage{
yQy(p, y, Qi, Q)
}
\arguments{
\item{p}{A permutation.}

\item{y}{A vector of the same length as \code{p}.}

\item{Qi}{A square matrix with row size the same as the length of \code{y}.}

\item{Q}{A square matrix with row size the same as the length of \code{y}.}
}
\description{
Compute \eqn{y'Qiy \ y'Qy} for a given permutation
}

