% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_load.r
\name{calc_load}
\alias{calc_load}
\title{Calculate loads}
\usage{
calc_load(
  ID,
  STP_treatment_steps,
  ID_next,
  inhabitants,
  STP_elimination = FALSE,
  compound_load_gramm_per_capita_and_day = NULL,
  compound_elimination_STP = NULL,
  compound_elimination_method = "micropollutants",
  with_lake_elimination = FALSE,
  lake_eliminination = NULL,
  add_absolute_load = FALSE,
  absolute_loads = FALSE,
  return_data = "loads"
)
}
\arguments{
\item{ID}{Character vector. Bla.}

\item{STP_treatment_steps}{Bla.}

\item{ID_next}{Bla.}

\item{inhabitants}{Bla.}

\item{STP_elimination}{Bla.}

\item{compound_load_gramm_per_capita_and_day}{Bla.}

\item{compound_elimination_STP}{Bla.}

\item{compound_elimination_method}{Bla.}

\item{with_lake_elimination}{Bla.}

\item{lake_eliminination}{Bla.}

\item{add_absolute_load}{Bla.}

\item{absolute_loads}{Bla.}

\item{return_data}{Bla.}
}
\value{
An integer vector the same length as `x`.
  `NA` strings have `NA` length.
}
\description{
Calculate loads
}
\examples{
Bla.

}
\seealso{
[wrap_table()]
}
