% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_table.r
\name{wrap_table}
\alias{wrap_table}
\title{Calculate ranges of compound loads and concentrations in a river network under different scenarios of sewage treatment.}
\usage{
wrap_table(
  input_table,
  STP_scenario_year = as.numeric(strsplit(as.character(Sys.Date()), "-")[[1]][1]),
  STP_reroute = TRUE,
  STP_filter_steps = TRUE,
  STP_discharge_per_capita = 400,
  compound_name = "compound_not_specified",
  compound_load_gramm_per_capita_and_day,
  compound_elimination_method = NULL,
  compound_elimination_STP = NULL,
  with_lake_elimination = FALSE,
  add_absolute_load = FALSE,
  use_columns_local_discharge = c("Q347_L_s_min", "Q347_L_s_max"),
  use_STP_elimination_rate = c("STP_elimination_min", "STP_elimination_max"),
  add_columns_from_input_table = NULL,
  scenario_name = compound_name,
  path_out = NULL,
  overwrite = TRUE,
  write_csv = TRUE,
  use_sep_csv = ","
)
}
\arguments{
\item{input_table}{Dataframe containing various information on sewage treatment plants and lakes, cp. example data \link{input_table}}

\item{STP_scenario_year}{4-digit integer value for the scenario year. Defaults to the current year.}

\item{STP_reroute}{Logical. Reroute STPs up to STP_scenario_year? Defaults to TRUE.}

\item{STP_filter_steps}{Logical. Filter STP treatment steps until a given STP_scenario_year? Defaults to TRUE.}

\item{STP_discharge_per_capita}{Single numeric value. Discharge per person and day \verb{[l / d]}.}

\item{compound_name}{Character string, name of the compound under consideration. Will be written into the exported Excel file, if such is used instead of csv.}

\item{compound_load_gramm_per_capita_and_day}{Vector of tow numeric value. Minimum and maximum amounts of the compound excreted per person and day \verb{[g / d]}.}

\item{compound_elimination_method}{Strings \code{"compound_specific"} or \code{"node_specific"} on how to calculate compound elimination, cp. Details.}

\item{compound_elimination_STP}{Required for compound_elimination_method \code{"compound_specific"}.
Dataframe with two rows for minium and maximum elimination fractions \verb{[0, 1]} for the following STP treatment steps aka column names:
\itemize{
\item COD_treatment
\item nitrification
\item denitrification
\item P_elimination
\item GAC
\item combi
\item ozonation
\item PAC
\item undefined: placeholder for yet unspecified but scheduled advanced treatment, e.g., a mean elimination fraction for GAC, combi, ozonation and PAC.
}}

\item{with_lake_elimination}{Logical. Include elimination by lakes? Defaults to TRUE.}

\item{add_absolute_load}{Logical, with default FALSE. Add further absolute loads at each node, and in additional to the ones calculated
via compound_load_gramm_per_capita_and_day plus compound_elimination_method?}

\item{use_columns_local_discharge}{Character vector with two column names from \code{input_table} for minimum and maximum discharge just downstream of an STP or lake. The defaults is \code{c("Q347_L_s_min", "Q347_L_s_max")}.}

\item{use_STP_elimination_rate}{Character vector with two column names from \code{input_table} for minimum and maximum STP-specific elimination rates,
required for compound_elimination_method \code{"node_specific"}. Defaults to \code{c("STP_elimination_min", "STP_elimination_max")}.}

\item{add_columns_from_input_table}{Character vector with names of columns from \code{input_table} to be attached and exported with the results table.}

\item{scenario_name}{Character string defaulting to \code{compound_name}. Name of the exported csv or Excel file path_out is not NULL.}

\item{path_out}{NULL (default) or string containing folder path for writing csv or Excel (not containing the file name, cp. \code{scenario_name}).}

\item{overwrite}{Logical, used if path_out is not NULL. Overwrite any existing file? Defaults to TRUE.}

\item{write_csv}{Logical, used if path_out is not NULL. Defaults to TRUE to write both exported files as csv, else exports one of these as Excel file (cp. return value section).}

\item{use_sep_csv}{String for csv sperataror, used if path_out is not NULL.}
}
\value{
This function either returns a dataframe (if \code{path_out} is not specified) or saves two file named \code{scenario_name_topo_matrix} and
\code{scenario_name} (or, if not provided, \code{compound_name}) into folder \code{path_out}.
The first is always a csv format and contains the topology (i.e. routing) matrix used for the calculations.
The second file is either a csv of Excel format (depending on \code{write_csv}) with the following result columns:
\itemize{
\item \code{ID}: Node IDs of STPs and lakes
\item \code{inhabitants_cumulated}: cumulated number of inhabitants just downstream of each STP or lake
\item \code{STP_local_discharge_L_s}: product of inhabitants and STP_discharge_per_capita from each STP \verb{[l / s]}
\item \code{STP_cumulated_discharge_L_s}: cumulated product of inhabitants and STP_discharge_per_capita just downstream each STP or lake node \verb{[l / s]}
\item \code{node_count_cumulated}: cumulated number of nodes upstream of a each STP or lake node
\item \code{Fraction_STP_discharge_of_river_local}: STP_local_discharge_L_s divided by the local discharge (i.e., first column in argument \code{use_columns_local_discharge})
\item \code{Fraction_STP_discharge_of_river_cumulated}: STP_cumulated_discharge_L_s divided by the local discharge (i.e., first column in argument \code{use_columns_local_discharge})
\item \code{Fraction_STP_discharge_without_advanced_treatment_of_river_cumulated}: local discharge (i.e., first column in argument \code{use_columns_local_discharge}) divided by
the sewage discharge from STPs without advanced treatments.
\item \code{Fraction_STP_discharge_of_river_local_includingSTPdischarge}: STP_local_discharge_L_s divided by the sum of local discharge (i.e., first column in argument
\code{use_columns_local_discharge}) and \code{STP_local_discharge_L_s}
\item \code{Fraction_STP_discharge_of_river_cumulated_includingSTPdischarge}: STP_cumulated_discharge_L_s divided by the sum of local discharge (i.e., first column in argument
\code{use_columns_local_discharge}) and \code{STP_local_discharge_L_s}
\item \code{Fraction_STP_discharge_without_advanced_treatment_of_river_cumulated_includingSTPdischarge}: STP_cumulated_discharge_L_s from STPs without advanced treatment divided by
the sum of local discharge (i.e., first column in argument \code{use_columns_local_discharge}) and \code{STP_local_discharge_L_s}
\item \code{Fraction_of_wastewater_only_C_removal}: fraction of cumulated STP discharge from STPs without nitrification, denitrification and any advanced treatment step, just downstream of each STP or lake.
\item \code{Fraction_of_wastewater_nitrification}: fraction of cumulated STP discharge from STPs with nitrification, but without denitrification and any advanced treatment step, just downstream of each STP or lake.
\item \code{Fraction_of_wastewater_denitrification}: fraction of cumulated STP discharge from STPs with nitrification and denitrification but without any advanced treatment step, just downstream of each STP or lake.
\item \code{Fraction_of_wastewater_advanced_treatment}: fraction of cumulated STP discharge from STPs with an advanced treatment step, just downstream of each STP or lake.
\item columns \code{input_load_local_g_d_max} and \code{input_load_local_g_d_min}: maximum and minimum compound amount released into each STP \verb{[g / d]}, i.e., inhabitants * compound_load_gramm_per_capita_and_day
\item columns \code{load_local_g_d_max} and \code{load_local_g_d_min}: maximum and minimum compound amount discharged from each STP after elimination \verb{[g / d]}
\item columns \code{load_cumulated_g_d_max} and \code{load_cumulated_g_d_min}: maximum and minimum cumulated compound amount just downstream of each STP or lake \verb{[g / d]}
\item columns \code{conc_local_ug_L_max} and \code{conc_local_ug_L_min}: \code{load_local_g_d_max} and \code{load_local_g_d_min} divided local discharge (i.e., relevant min/max columns in argument \code{use_columns_local_discharge})
\item columns \code{conc_cumulated_ug_L_max} and \code{conc_cumulated_ug_L_min}: \code{load_cumulated_g_d_max} and \code{load_cumulated_g_d_min} divided local discharge (i.e., relevant min/max columns in argument \code{use_columns_local_discharge})
}

In case of the Excel file output, the first rows in the spreadsheet also contain the parameters and the \code{compound_name} used.
}
\description{
Wrapper function on \code{\link[=calc_load]{calc_load()}} to calculate minimum and maximum loads and concentrations of a compound in a river network from an input table
containing information on sewage treatment plants (STPs) and lake nodes. \code{STP_scenario_year} allows simulations for different stages of advanced treatment
or a rerouting of STPs.
}
\details{
Check function \code{\link[=calc_load]{calc_load()}} for the underlying approach on calculating loads from human compound inputs and their elimination in STPs (cp. argument \code{compound_elimination_method})
and lakes (cp. \code{with_lake_elimination}).
Using \link{table_input}, this wrapper extends this approach by estimating minimum and maximum loads and their aquatic concentrations.
By using an \code{STP_scenario_year} and column \code{starting_year_advanced_treatment} in \link{input_table} (and possibly in combination with \code{STP_reroute}) data for different points in
time can be simulated.
}
\note{
In contrast to column \code{type_advanced_treatment} of argument \code{STP_treatment_steps} in function \code{\link[=calc_load]{calc_load()}},
the same column in \code{input_table} may also contain the entry \code{"redirection"}, apart from \code{"GAC"}, \code{"combi"}, \code{"ozonation"}, \code{"PAC"} or \code{"undefined"}.
It redirects the influent from one STP to another \code{redirecting_STP_target_STP_ID} (another column in \link{input_table}) when \code{STP_reroute} is set to TRUE, in combination
with argument \code{STP_scenario_year} and \code{starting_year_advanced_treatment} (yet another column in \link{input_table}) .
}
\examples{

compound_load_gramm_per_capita_and_day <- c(100 * 1E-6,	500 * 1E-6)

compound_elimination_STP <- data.frame(	
	COD_treatment = c(0.4, 0.5),
	nitrification = c(0.3, 0.6),
	denitrification = c(0.15, 0.2),
	P_elimination = c(0.1, 0.2),
	GAC = c(0, 0.15),
	combi = c(0, 0.05),
	ozonation = c(0.4, 0.7),
	PAC = c(0, 0.3),
	undefined = c(0, 0.12)
)

results <- wrap_table(
	input_table = input_table,
	STP_scenario_year = 2030,
	STP_reroute = TRUE,
	STP_filter_steps = TRUE,
	STP_discharge_per_capita = 400,
	compound_name = "Diclofenac",
	compound_load_gramm_per_capita_and_day = compound_load_gramm_per_capita_and_day,
	compound_elimination_method = "compound_specific",
	compound_elimination_STP = compound_elimination_STP,
	with_lake_elimination = TRUE,
	add_absolute_load = TRUE,
	use_columns_local_discharge = c("Q347_L_s_min", "Q347_L_s_max"),
	use_STP_elimination_rate = c("STP_elimination_min", "STP_elimination_max"),
	add_columns_from_input_table = c("ID_next", "X_position", "Y_position"),
	scenario_name = compound_name,	
	path_out = NULL
)


# Above [input_table] is lazy-loaded for this package.
# In addition, an Excel version model_inputs.xlsx of [input_table] exists in the extdata folder of this package, to be found under:
list.files(system.file("extdata", package = "SSM"))
# The first two rows contain comments, and must be skipped when loading via:
xlsxFile_path <- file.path( system.file("extdata", package = "SSM"), "model_inputs.xlsx")
input_table <- openxlsx:::read.xlsx(xlsxFile = xlsxFile_path, sheet = "model_inputs", startRow = 3)


}
\seealso{
\code{\link[=calc_load]{calc_load()}}, \link{input_table}
}
